requirements.txt
datasets==3.6.0
fastapi==0.116.1
ipython==9.1.0
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.75
langchain_google_genai==2.1.10
langchain_ollama==0.3.7
langchain_postgres==0.0.15
langchain_text_splitters==0.3.11
langgraph==0.6.6
matplotlib==3.10.6
Pillow==11.3.0
protobuf==6.32.0
pydantic==2.11.7
pymupdf==1.26.4
python-dotenv==1.1.1
ragas==0.3.2
Requests==2.32.5
streamlit==1.49.1
tqdm==4.67.1
typing_extensions==4.15.0
